chore: move most logic to shared partial
